% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot FARS data on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing the state number.}

\item{year}{An integer representing the year.}
}
\value{
A map with the plotted FARS data for the given state and year, or an invisible NULL if there are no accidents to plot.
}
\description{
This function plots FARS data on a map for a given state and year.
}
\details{
If the state number is not valid, the function stops with an error message. If there are no accidents to plot, a message is displayed and NULL is returned.
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}
}
