% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to the file.}
}
\value{
A tibble containing the FARS data from the file.
}
\description{
This function reads a CSV file containing FARS data and returns it as a tibble.
}
\details{
If the file does not exist, the function will stop with an error message.
}
\examples{
\dontrun{
fars_data <- fars_read("AER.csv")
}
}
