% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of integers representing the years.}
}
\value{
A list of tibbles containing the FARS data for each year.
}
\description{
This function reads FARS data for multiple years and returns a list of tibbles.
}
\details{
If a year does not correspond to a valid file, a warning is issued and NULL is returned for that year.
}
\examples{
\dontrun{
data_list <- fars_read_years(c(2013, 2014, 2015))
}
}
