% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS data by year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of integers representing the years.}
}
\value{
A tibble summarizing the number of accidents by year and month.
}
\description{
This function summarizes FARS data by year and month.
}
\details{
The function binds the yearly data into a single data frame, groups by year and month, and then summarizes the number of accidents.
}
\examples{
\dontrun{
summary <- fars_summarize_years(c(2013, 2014, 2015))
}
}
