% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{make_filename}
\alias{make_filename}
\title{Create a FARS filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or a value that can be coerced to an integer, representing the year.}
}
\value{
A character string with the FARS filename for the given year.
}
\description{
This function creates a FARS filename for a given year.
}
\examples{
make_filename(2013)
}
